refactor(team): extract subscribe/streams into broker_streams.go
Slice 13 (final). Moves the subscriber-fan-out, agent-stream buffer,
broadcast-dedupe metric, and activity watchdog (~423 lines) into a new
broker_streams.go:
- agentStreamBuffer type + Push/subscribe/recent
- AgentStream method
- SubscribeMessages, SubscribeActions, SubscribeActivity
- officeChangeEvent type + SubscribeOfficeChanges + publishOfficeChangeLocked
- SubscribeWikiEvents, PublishWikiEvent
- WikiWorker, WikiIndex getters
- UpdateAgentActivity
- duplicateBroadcastWindow + duplicateBroadcastSimilarity constants
- isDuplicateAgentBroadcastLocked, normalizeBroadcastContent,
  jaccardWordSimilarity, uniqueWordSet
- activityWatchdogEnabled var + staleActivityThreshold const
- reapStaleActivityLocked, runActivityWatchdog

publish*Locked and appendMessageLocked stay in broker.go — central
plumbing called from many siblings; a separate dedicated PR is safer.

Relocates 5 stream tests from broker_test.go into broker_streams_test.go:
TestBrokerMessageSubscribersReceivePostedMessages,
TestBrokerActionSubscribersReceiveTaskLifecycle,
TestReapStaleActivityLocked, TestBrokerActivitySubscribersReceiveUpdates,
TestBrokerEventsEndpointStreamsMessages.

Adds three characterization tests:
- TestAgentStreamBuffer_RecentReturnsBoundedHistory (cap=2000, trailing
  window preserved, subscribers see live writes)
- TestJaccardWordSimilarity_KnownPairs (identity, disjoint,
  empty-short-circuit, near-paraphrase)
- TestUniqueWordSet_StripsPunctuation (trailing punctuation collapses,
  intra-word apostrophes preserved)

Test suite passes under -race. Total package coverage: 63.3% (vs 62.9%
baseline floor).
