+++ internal/team/broker_test.go
@@ -1,7 +1,6 @@
 package team
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"encoding/json"
@@ -229,182 +228,10 @@ func TestBrokerLoadsLastGoodSnapshotWhenPrimaryStateIsClobbered(t *testing.T) {
 	}
 }
 
-func TestBrokerMessageSubscribersReceivePostedMessages(t *testing.T) {
-	b := newTestBroker(t)
-	msgs, unsubscribe := b.SubscribeMessages(4)
-	defer unsubscribe()
-
-	want, err := b.PostMessage("ceo", "general", "Push this immediately", nil, "")
-	if err != nil {
-		t.Fatalf("PostMessage: %v", err)
-	}
-
-	select {
-	case got := <-msgs:
-		if got.ID != want.ID || got.Content != want.Content {
-			t.Fatalf("unexpected subscribed message: %+v", got)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for subscribed message")
-	}
-}
-
-func TestBrokerActionSubscribersReceiveTaskLifecycle(t *testing.T) {
-	b := newTestBroker(t)
-	actions, unsubscribe := b.SubscribeActions(4)
-	defer unsubscribe()
-
-	if _, _, err := b.EnsureTask("general", "Landing page", "Build the hero", "fe", "ceo", ""); err != nil {
-		t.Fatalf("EnsureTask: %v", err)
-	}
-
-	select {
-	case got := <-actions:
-		if got.Kind != "task_created" {
-			t.Fatalf("expected task_created action, got %+v", got)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for subscribed action")
-	}
-}
-
-func TestReapStaleActivityLocked(t *testing.T) {
-	b := newTestBroker(t)
-	now := time.Now().UTC()
-	stale := now.Add(-10 * time.Minute).Format(time.RFC3339)
-	fresh := now.Add(-1 * time.Minute).Format(time.RFC3339)
-
-	b.activity = map[string]agentActivitySnapshot{
-		"stale-active":   {Slug: "stale-active", Status: "active", Activity: "tool_use", LastTime: stale},
-		"stale-thinking": {Slug: "stale-thinking", Status: "thinking", Activity: "thinking", LastTime: stale},
-		"fresh-active":   {Slug: "fresh-active", Status: "active", Activity: "tool_use", LastTime: fresh},
-		"already-idle":   {Slug: "already-idle", Status: "idle", Activity: "idle", LastTime: stale},
-		"already-error":  {Slug: "already-error", Status: "error", Activity: "error", LastTime: stale},
-		"bad-time":       {Slug: "bad-time", Status: "active", Activity: "tool_use", LastTime: "not-a-time"},
-	}
-
-	b.mu.Lock()
-	reset := b.reapStaleActivityLocked(now)
-	b.mu.Unlock()
-
-	if len(reset) != 2 {
-		t.Fatalf("expected 2 stale agents reaped, got %d: %+v", len(reset), reset)
-	}
-	for _, snap := range reset {
-		if snap.Status != "idle" {
-			t.Errorf("reaped agent %q should be idle, got %q", snap.Slug, snap.Status)
-		}
-		if snap.Slug != "stale-active" && snap.Slug != "stale-thinking" {
-			t.Errorf("unexpected reaped slug: %q", snap.Slug)
-		}
-	}
-
-	if b.activity["fresh-active"].Status != "active" {
-		t.Error("fresh-active should not be reaped")
-	}
-	if b.activity["already-idle"].Status != "idle" {
-		t.Error("already-idle should be unchanged")
-	}
-	if b.activity["already-error"].Status != "error" {
-		t.Error("already-error should be unchanged")
-	}
-	if b.activity["bad-time"].Status != "active" {
-		t.Error("unparseable LastTime should be left alone")
-	}
-}
-
 func TestBrokerStopIsIdempotent(t *testing.T) {
 	b := newTestBroker(t)
 	b.Stop()
 	b.Stop()
-}
-
-func TestBrokerActivitySubscribersReceiveUpdates(t *testing.T) {
-	b := newTestBroker(t)
-	updates, unsubscribe := b.SubscribeActivity(4)
-	defer unsubscribe()
-
-	b.UpdateAgentActivity(agentActivitySnapshot{
-		Slug:     "ceo",
-		Status:   "active",
-		Activity: "tool_use",
-		Detail:   "running rg",
-		LastTime: time.Now().UTC().Format(time.RFC3339),
-	})
-
-	select {
-	case got := <-updates:
-		if got.Slug != "ceo" || got.Activity != "tool_use" || got.Detail != "running rg" {
-			t.Fatalf("unexpected activity update: %+v", got)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for subscribed activity")
-	}
-}
-
-func TestBrokerEventsEndpointStreamsMessages(t *testing.T) {
-	b := newTestBroker(t)
-	b.channels = []teamChannel{
-		{
-			Slug:    "general",
-			Name:    "general",
-			Members: []string{"operator"},
-		},
-		{
-			Slug:    "planning",
-			Name:    "planning",
-			Members: []string{"operator", "planner"},
-		},
-	}
-	if err := b.StartOnPort(0); err != nil {
-		t.Fatalf("failed to start broker: %v", err)
-	}
-	defer b.Stop()
-
-	base := fmt.Sprintf("http://%s", b.Addr())
-	req, _ := http.NewRequest(http.MethodGet, base+"/events?token="+b.Token(), nil)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("open event stream: %v", err)
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		raw, _ := io.ReadAll(resp.Body)
-		t.Fatalf("expected 200 opening event stream, got %d: %s", resp.StatusCode, raw)
-	}
-
-	lines := make(chan string, 16)
-	go func() {
-		scanner := bufio.NewScanner(resp.Body)
-		for scanner.Scan() {
-			lines <- scanner.Text()
-		}
-		close(lines)
-	}()
-
-	if _, err := b.PostMessage("ceo", "general", "Stream this", nil, ""); err != nil {
-		t.Fatalf("PostMessage: %v", err)
-	}
-
-	deadline := time.After(2 * time.Second)
-	var sawEvent bool
-	var sawPayload bool
-	for !(sawEvent && sawPayload) {
-		select {
-		case line, ok := <-lines:
-			if !ok {
-				t.Fatal("event stream closed before receiving message")
-			}
-			if strings.Contains(line, "event: message") {
-				sawEvent = true
-			}
-			if strings.Contains(line, `"content":"Stream this"`) {
-				sawPayload = true
-			}
-		case <-deadline:
-			t.Fatalf("timed out waiting for message event (event=%v payload=%v)", sawEvent, sawPayload)
-		}
-	}
 }
 
 func TestBrokerPersistsNotificationCursorWithoutMessages(t *testing.T) {
